+++ rana_qgis_plugin/rana_qgis_plugin.py
@@ -242,6 +242,9 @@ class RanaQgisPlugin:
             self.loader.loading_cancelled.connect(self.rana_browser.enable)
             self.rana_browser.download_file_selected.connect(self.loader.download_file)
             self.rana_browser.download_results_selected.connect(
+                self.rana_browser.disable
+            )
+            self.rana_browser.download_results_selected.connect(
                 self.loader.download_results
             )
             self.rana_browser.open_in_qgis_selected.connect(self.rana_browser.disable)
@@ -249,16 +252,21 @@ class RanaQgisPlugin:
             self.rana_browser.save_vector_styling_selected.connect(
                 self.rana_browser.disable
             )
+            self.loader.download_results_cancelled.connect(self.rana_browser.enable)
             self.rana_browser.download_file_selected.connect(self.rana_browser.disable)
-            self.rana_browser.download_results_selected.connect(
-                self.rana_browser.disable
-            )
             self.rana_browser.start_simulation_selected.connect(
                 self.rana_browser.disable
             )
             self.rana_browser.start_simulation_selected.connect(
                 self.loader.start_simulation
             )
+            self.rana_browser.start_simulation_selected_with_revision.connect(
+                self.rana_browser.disable
+            )
+            self.rana_browser.start_simulation_selected_with_revision.connect(
+                self.loader.start_simulation
+            )
+            self.rana_browser.delete_file_selected.connect(self.loader.delete_file)
 
             self.loader.file_download_finished.connect(self.rana_browser.enable)
             self.loader.file_download_failed.connect(self.rana_browser.enable)
@@ -274,6 +282,7 @@ class RanaQgisPlugin:
             self.loader.simulation_started.connect(self.rana_browser.enable)
             self.loader.simulation_cancelled.connect(self.rana_browser.enable)
             self.loader.simulation_started_failed.connect(self.rana_browser.enable)
+            self.loader.file_deleted.connect(self.rana_browser.refresh)
 
         self.iface.addTabifiedDockWidget(
             Qt.DockWidgetArea.RightDockWidgetArea, self.dock_widget, raiseTab=True