+++ HISTORY.rst
@@ -4,7 +4,9 @@ History
 1.1.21 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Add revision view and new simulation button
+- Replace file interaction buttons with context menu
+- Download generated postprocessing rasters.
 
 
 1.1.20 (2025-11-18)